Add SkipRunes, DropRunes (#6)
Per discussion in #3, I've implemented `SkipRunes`, `DropRunes`.

I have also made the following changes:

- Optimised `DropString`
- Previously, I allocated a new byte buffer and then cast it to a
string, which allocates again. It was possible to cast it using `unsafe`
and avoid the extra allocation, but at the time I figured it wasn't
worth it as . Since then I have found `strings.Builder`, which
internally performs the `unsafe` conversion and saves us an allocation.
- Removed `append` in `Drop`
- We know the length of the slice we are returning, but we set the
length to 0 in the make call. It was overwritten by `append`, but we can
eliminate that by just making the slice with the final length at the
beginning and copying the elements with `copy`. Should have no
(negative) change in performance as all allocations in Go are zero'd
anyways (so no internal optimisations done by `append`).
- Clarify doc comments
  - Clarified doc comments as discussed in #3.
- Remove unnecessary allocation in `Skip`, `SkipString`
- We previously allocated a zero value when we skip more than the number
of elements in the slice, possibly dropping the underlying array of the
slice passed in to the function. This is the intended behaviour of
`Drop`, but not `Skip`, and so now we return a slice of the same
underlying array, but with length 0, maintaining a reference to that
underlying array.
- Split rune functions into separate file
- I believe that with the project size increasing we should start
splitting up new functions into different files for easier development.
- Restrict `Skip`,`Drop`,... to non-negative integers
- I realised that we don't actually have defined behaviour for when a
user passes in a negative integer, but instead of adding checks and
documenting, I reasoned that it would just be best to restrict this via
the type system. The reason we don't want to accept negative integers is
because we already mentioned an `Unskip` method, and it is up to debate
whether we would ever expose such functionality (re. idiomaticity)
- Separate `String` functions into their own file
  - Same reason why `Runes` functions are in their own file.